+++ CHANGELOG.rst
@@ -28,6 +28,10 @@ Deprecations:
 * The ``from_override_rules`` method of ``PageObjectRegistry`` is now deprecated.
   Use ``from_apply_rules`` instead.
 * The ``web_poet.overrides`` is deprecated. Use ``web_poet.rules`` instead.
+* The ``PageObjectRegistry.get_overrides`` is deprecated.
+  Use ``PageObjectRegistry.get_rules`` instead.
+* The ``PageObjectRegistry.search_overrides`` is deprecated.
+  Use ``PageObjectRegistry.search_rules`` instead.
 
 0.5.1 (2022-09-23)
 ------------------